+++ rust/processor/src/db/postgres/migrations/2024-10-20-152751_rework_price_feed/down.sql
@@ -1,6 +1,6 @@
 -- This file should undo anything in `up.sql`
 CREATE OR REPLACE FUNCTION price_feed() RETURNS TABLE(
-  market_id BIGINT,
+  market_id NUMERIC,
   symbol_bytes BYTEA,
   symbol_emojis TEXT[],
   market_address VARCHAR(66),
